refactor(web): extract TimePickerButton
Pull the day-anchor picker (button trigger, vue-datepicker setup,
locale-aware 12/24-hour, themed popup) out of SkeletonSection into
a reusable component so the editable Review tab can use the same
control without duplicating ~70 lines of CSS and config.
